email_triage_env/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "email-triage-env"
version = "0.1.0"
description = "Email Triage OpenEnv environment — classify, prioritize, and reply to support emails"
requires-python = ">=3.11"
dependencies = [
    "openenv-core>=0.2.2",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "uvicorn[standard]>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "httpx>=0.25.0",
]

[project.scripts]
server = "email_triage_env.server.app:main"

[tool.setuptools.packages.find]
include = ["email_triage_env*"]
